Fixed some relationshipts

+++ src/main/java/org/themarioga/cclh/commons/services/impl/GameServiceImpl.java
@@ -1,5 +1,6 @@
 package org.themarioga.cclh.commons.services.impl;
 
+import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +10,10 @@ import org.themarioga.cclh.commons.enums.CardTypeEnum;
 import org.themarioga.cclh.commons.enums.ErrorEnum;
 import org.themarioga.cclh.commons.enums.GameStatusEnum;
 import org.themarioga.cclh.commons.enums.GameTypeEnum;
+import org.themarioga.cclh.commons.exceptions.ApplicationException;
+import org.themarioga.cclh.commons.exceptions.dictionary.DictionaryDoesntExistsException;
 import org.themarioga.cclh.commons.exceptions.game.*;
+import org.themarioga.cclh.commons.exceptions.room.RoomDoesntExistsException;
 import org.themarioga.cclh.commons.models.*;
 import org.themarioga.cclh.commons.services.intf.*;
 import org.themarioga.cclh.commons.util.Assert;
@@ -43,36 +47,48 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
-    public Game create(long roomId, String roomName, long ownerId, long creatorId) {
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
+    public Game create(long roomId, String roomName, long roomOwnerId, long creatorId) {
         logger.debug("Creating game in room: {}({})", roomId, roomName);
 
         // Create or load room
-        Room room = roomService.createOrReactivate(roomId, roomName, ownerId);
+        Room room = roomService.createOrReactivate(roomId, roomName, roomOwnerId);
 
         // Create game
         Game game = new Game();
         game.setRoom(room);
         game.setCreator(userService.getById(creatorId));
+        game.setStatus(GameStatusEnum.CREATED);
 
         return gameDao.create(game);
     }
 
     @Override
-    public void delete(Game game) {
-        logger.debug("Deleting game: {}", game);
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
+    public void delete(long roomId) {
+        logger.debug("Deleting game: {}", roomId);
+
+        Room room = roomService.getById(roomId);
+
+        if (room == null) throw new RoomDoesntExistsException(roomId);
+
+        Game game = gameDao.getByRoom(room);
+
+        if (game == null) throw new GameDoesntExistsException(room.getId());
 
         gameDao.delete(game);
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game setType(Game game, GameTypeEnum type) {
         logger.debug("Setting type {} to game {}", type, game);
 
         // Check game exists
         Assert.assertNotNull(game, ErrorEnum.GAME_NOT_FOUND);
 
-        // Check if this query has already been done
-        if (game.getType() != null) throw new GameAlreadyConfiguredException(game.getId());
+        // Check if the game has already started
+        if (game.getStatus() == GameStatusEnum.STARTED) throw new GameAlreadyStartedException(game.getId());
 
         // Set the type to game
         game.setType(type);
@@ -81,14 +97,15 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game setNumberOfCardsToWin(Game game, int numberOfCards) {
         logger.debug("Setting number of cards {} to game {}", numberOfCards, game);
 
         // Check game exists
         Assert.assertNotNull(game, ErrorEnum.GAME_NOT_FOUND);
 
-        // Check if this query has already been done
-        if (game.getNumberOfCardsToWin() != null) throw new GameAlreadyConfiguredException(game.getId());
+        // Check if the game has already started
+        if (game.getStatus() == GameStatusEnum.STARTED) throw new GameAlreadyStartedException(game.getId());
 
         // Set the number of cards to win
         game.setNumberOfCardsToWin(numberOfCards);
@@ -97,14 +114,15 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game setMaxNumberOfPlayers(Game game, int maxNumberOfPlayers) {
         logger.debug("Setting max number of players {} to game {}", maxNumberOfPlayers, game);
 
         // Check game exists
         Assert.assertNotNull(game, ErrorEnum.GAME_NOT_FOUND);
 
-        // Check if this query has already been done
-        if (game.getMaxNumberOfPlayers() != null) throw new GameAlreadyConfiguredException(game.getId());
+        // Check if the game has already started
+        if (game.getStatus() == GameStatusEnum.STARTED) throw new GameAlreadyStartedException(game.getId());
 
         // Set the max number of players
         game.setMaxNumberOfPlayers(maxNumberOfPlayers);
@@ -113,15 +131,21 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game setDictionary(Game game, long dictionaryId) {
         logger.debug("Setting dictionary {} to game {}", dictionaryId, game);
 
         // Check game exists
         Assert.assertNotNull(game, ErrorEnum.GAME_NOT_FOUND);
 
-        // Check if this query has already been done
+        // Check if the game has already started
+        if (game.getStatus() == GameStatusEnum.STARTED) throw new GameAlreadyStartedException(game.getId());
+
+        // Find the dictionary
         Dictionary dictionary = dictionaryService.findOne(dictionaryId);
-        if (dictionary == null) throw new GameAlreadyConfiguredException(game.getId());
+
+        // Check if the dictionary exists
+        if (dictionary == null) throw new DictionaryDoesntExistsException(dictionaryId);
 
         // Set the dictionary
         game.setDictionary(dictionary);
@@ -130,6 +154,7 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game addPlayer(Game game, User user) {
         logger.debug("Creating player from user {} in game {}", user, game);
 
@@ -149,6 +174,7 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game startGame(Game game) {
         logger.debug("Starting game {}", game);
 
@@ -177,6 +203,7 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game startRound(Game game) {
         logger.debug("Starting round for game {}", game);
 
@@ -195,6 +222,7 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public Game endRound(Game game) {
         logger.debug("Ending round for game {}", game);
 
@@ -208,6 +236,7 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = ApplicationException.class)
     public void voteForDeletion(Game game, Player player) {
         logger.debug("Player {} vote for the deletion of the game {}", player, game);
 
@@ -217,14 +246,23 @@ public class GameServiceImpl implements GameService {
         // Check player exists
         Assert.assertNotNull(player, ErrorEnum.GAME_NOT_FOUND);
 
-
+        // ToDo: make this
     }
 
     @Override
-    public Game getByRoomId(Room room) {
+    @Transactional(value = Transactional.TxType.SUPPORTS, rollbackOn = ApplicationException.class)
+    public Game getByRoom(Room room) {
         logger.debug("Getting game with room: {}", room);
 
-        return gameDao.getByRoomId(room);
+        return gameDao.getByRoom(room);
+    }
+
+    @Override
+    @Transactional(value = Transactional.TxType.SUPPORTS, rollbackOn = ApplicationException.class)
+    public Game getByRoomId(long roomId) {
+        logger.debug("Getting game with room id: {}", roomId);
+
+        return gameDao.getByRoom(roomService.getById(roomId));
     }
 
     private void addBlackCardsToTableDeck(Game game) {
